refactor: move schema to correct folder and validation error
