+++ services/credential-issuance-server/src/apis/schema.api.ts
@@ -1,11 +1,11 @@
 import { Request, Response } from "express";
-import { SCHEMA_ACDC } from "../utils/contexts/schemaAcdc";
+import { SCHEMA_ACDC } from "../utils/schemas/schemaAcdc";
 
 async function schemaApi(req: Request, res: Response) {
   const { id } = req.params;
   const data = SCHEMA_ACDC[id];
   if (!data) {
-    return res.status(404).send("Schema said is not valid");
+    return res.status(404).send("Schema for given SAID not found");
   }
   return res.send(data);
 }
